Рефакторинг и борьба с html формой в jsp и json отправкой из формы

+++ src/main/java/io/khasang/moika/service/impl/CarServiceImpl.java
@@ -51,9 +51,8 @@ public class CarServiceImpl implements CarService {
 
     @Override
     public void deleteCar(long id) {
-        Car car = new Car();
-        car.setId(id);
-        carDao.delete(car);
+        Car car = carDao.get(id);
+        if (car != null)  carDao.delete(car);
     }
 
     @Override
